products/tier3/rag-avancado/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# LangChain & RAG
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2
chromadb==0.4.22

# LLM
openai==1.3.9

# Embeddings
sentence-transformers==2.2.2
tiktoken==0.5.2

# Vector Store
faiss-cpu==1.7.4

# Utils
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-dateutil==2.8.2
pytz==2023.3

# WebSocket
websockets==12.0

# Monitoring
python-json-logger==2.0.7

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1

